feat #20: 인증 정보 팝업 -> 화면 변경
+ UIImageViewExt, loadImage 추가
