+++ dogether/Domain/Entity/Enum/PopupTypes.swift
@@ -10,7 +10,6 @@ import UIKit
 enum PopupTypes {
     case alert
     case certification
-    case certificationInfo
     case rejectReason
 }
